ref(new-widget-builder-experience): Merge MetricsSearchbar component into ReleaseSearchBar
